+++ internal/model/price.go
@@ -1,19 +1,19 @@
 package model
 
 type Request struct {
-	VenueSlug  string `json:"venue_slug"`
-	CartValue  int    `json:"cart_value"`
+	VenueSlug  string  `json:"venue_slug"`
+	CartValue  float64 `json:"cart_value"`
 	UserCoords Location
 }
 
 type Response struct {
-	TotalPrice     int `json:"total_price"`
-	OrderSurcharge int `json:"small_order_surcharge"`
-	CartValue      int `json:"cart_value"`
+	TotalPrice     float64 `json:"total_price"`
+	OrderSurcharge float64 `json:"small_order_surcharge"`
+	CartValue      float64 `json:"cart_value"`
 	Delivery       Delivery
 }
 
 type Delivery struct {
-	Fee      int `json:"fee"`
-	Distance int `json:"distance"`
+	Fee      float64 `json:"fee"`
+	Distance float64 `json:"distance"`
 }
